+++ ofht_map.go
@@ -1159,5 +1159,5 @@ func ofhtIntHash(x uintptr) uint32 {
 	if ofhtUseRawIntHash {
 		return uint32(x ^ (x >> 32))
 	}
-	return uint32((x * opt.HashPrime) >> 32)
+	return uint32((uint64(x) * opt.HashPrime) >> 32)
 }
